// client/src/initialState.ts
export type EncryptionPathPair = {
    certificate_path: string;
    private_key_path: string;
};

export type EncryptionPathPairStatus = EncryptionPathPair & {
    issuer: string;
    key_type: string;
    not_after: string;
    not_before: string;
    subject: string;
    warning_validation: string;
    dns_names: string[];
    valid_chain: boolean;
    valid_key: boolean;
    valid_cert: boolean;
    valid_pair: boolean;
};

export type EncryptionData = {
    processing: boolean;
    processingConfig: boolean;
    processingValidate: boolean;
    can_apply: boolean;
    panel_can_apply: boolean;
    enabled: boolean;
    serve_plain_dns: boolean;
    dns_names: any;
    panel_dns_names: any;
    force_https: boolean;
    issuer: string;
    panel_issuer: string;
    key_type: string;
    panel_key_type: string;
    not_after: string;
    panel_not_after: string;
    not_before: string;
    panel_not_before: string;
    panel_server_name?: string;
    panel_server_url_path?: string;
    panel_server_port?: number;
    port_dns_over_tls?: number;
    port_dns_over_quic?: number;
    port_https?: number;
    port_dnscrypt?: number;
    subject: string;
    panel_subject: string;
    valid_chain: boolean;
    panel_valid_chain: boolean;
    valid_key: boolean;
    panel_valid_key: boolean;
    valid_cert: boolean;
    panel_valid_cert: boolean;
    valid_pair: boolean;
    panel_valid_pair: boolean;
    status_cert: string;
    status_key: string;
    private_key: string;
    panel_private_key: string;
    server_name: string;
    warning_validation: string;
    panel_warning_validation: string;
    certificate_chain: string;
    panel_certificate_chain: string;
    certificate_path: string;
    panel_certificate_path: string;
    certificate_key_pairs: EncryptionPathPair[];
    panel_certificate_key_pairs: EncryptionPathPair[];
    certificate_key_pair_statuses: EncryptionPathPairStatus[];
    panel_certificate_key_pair_statuses: EncryptionPathPairStatus[];
    private_key_path: string;
    panel_private_key_path: string;
    private_key_saved: boolean;
    panel_private_key_saved: boolean;
    dns_assigned_certificate_ids: number[];
    panel_assigned_certificate_ids: number[];
    dns_over_quic_url_path?: string;
    allow_unencrypted_doh?: boolean;
    dnscrypt_config_file?: string;
};